Renames City to CityDetails
